# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SpendCapUpdate": {
        "additionalProperties": false,
        "properties": {
          "spend_cap_cents": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spend Cap Cents"
          }
        },
        "title": "SpendCapUpdate",
        "type": "object"
      },
...
      "UsageSummary": {
        "additionalProperties": false,
        "properties": {
          "billing_status": {
            "title": "Billing Status",
            "type": "string"
          },
          "completed_decisions": {
            "title": "Completed Decisions",
            "type": "integer"
          },
          "included_decisions": {
            "title": "Included Decisions",
            "type": "integer"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "overage_decisions": {
            "title": "Overage Decisions",
            "type": "integer"
          },
          "payment_method_present": {
            "title": "Payment Method Present",
            "type": "boolean"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "projected_overage_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Overage Cents"
          },
          "remaining_free_decisions": {
            "title": "Remaining Free Decisions",
            "type": "integer"
          },
          "spend_cap_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spend Cap Cents"
          },
          "stripe_projection_is_async": {
            "default": true,
            "title": "Stripe Projection Is Async",
            "type": "boolean"
          }
        },
        "required": [
          "organization_id",
          "period",
          "completed_decisions",
          "included_decisions",
          "overage_decisions",
          "remaining_free_decisions",
          "billing_status",
          "payment_method_present",
          "spend_cap_cents",
          "projected_overage_cents"
        ],
        "title": "UsageSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentMesh Risk-Aware Support API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/usage": {
      "get": {
        "operationId": "organization_usage_api_v1_organizations__organization_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          },
          {
            "in": "header",
            "name": "x-dev-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Email"
            }
          },
          {
            "in": "header",
            "name": "x-dev-aal",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Organization Usage",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/usage/spend-cap": {
      "patch": {
        "operationId": "set_spend_cap_api_v1_organizations__organization_id__usage_spend_cap_patch",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          },
          {
            "in": "header",
            "name": "x-dev-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Email"
            }
          },
          {
            "in": "header",
            "name": "x-dev-aal",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Aal"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpendCapUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Spend Cap",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}